User Handling, Routing, & Profile
Redid the way users and their context is handled.  Instead of pulling
the context for a user on every component, I properly pass the
original user context I pull in App.js to wherever it is needed.

Using Routing Me is now protected properly, and can't be accessed unless
the user is signed in.

Started adding profile aspects like Followers and Following, and a
overall profile management page is ongoing.